refactor(colorutils): move color utilities to modules/color
