+++ src/modules/color/index.js
@@ -1,0 +1,11 @@
+export {
+  clamp360,
+  rotateHue,
+  extractHSLColorComponents,
+  toHSL,
+} from './hsl'
+
+export {
+  extractRGBAColorComponents,
+  toRGBA,
+} from './rgba'
